Pendulum project report
Files related to my phy324 experiment design project. This project asked that I study the motion of a planar pendulum using an apparatus designed at home. The purpose of this project was to design a simple experiment from the ground up. This experiment helped me learn how to optimize my tools to minimize uncertainty.